requirements.txt
google-generativeai>=0.8.0
yfinance>=0.2.36
feedparser>=6.0.11
requests>=2.31.0
python-dateutil>=2.8.2
resend>=0.8.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
